6-size.c

+++ 0x00-hello_world/6-size.c
@@ -7,10 +7,10 @@
  */
 int main(void)
 {
-	printf("Size of a char: %lu byte(s)\n", sizeof(char));
-	printf("Size of an int: %lu byte(s)\n", sizeof(int));
-	printf("Size of an long int: %lu byte(s)\n", sizeof(long int));
-	printf("Size of an long long int: %lu byte(s)\n", sizeof(long long int));
-	printf("Size of an float: %lu byte(s)\n", sizeof(float));
+	printf("Size of a char: %i byte(s)\n", (int) sizeof(char));
+	printf("Size of an int: %i byte(s)\n", (int) sizeof(int));
+	printf("Size of an long int: %i byte(s)\n", (int) sizeof(long int));
+	printf("Size of an long long int: %i byte(s)\n", (int) sizeof(long long int));
+	printf("Size of an float: %i byte(s)\n", (int) sizeof(float));
 	return (0);
 }
